requirements.txt
openai>=1.0.0
requests>=2.28.0
tavily-python>=0.3.0
